requirements.txt
Flask==2.3.2
openai==0.27.0
pinecone-client==2.2.1
python-dotenv==1.0.0
